Add option to authenticate gRPC requests
To be able to talk to gRPC endpoints with SSL or authentication, we
need to call a function that generates the SSL cert and/or any other
authentication metadata.

Add a new attribute to the gRPC handler that specifies a path to a
function that generates the needed authentication data. This is then
called and passed to the gRPC client to be used to authenticate calls.