Add option to specify grpc Metadata and tighten up typing (#519)
* add option for adding grpc MetaData

Add ability to either specify static or dynamic MetaData from the
context in the config.

Static MetaData is read directly and stringfied, if it is a non string
value.

Dynamic MetaData is extracted from the GraphQL context via Lodash. This
means that specifing deeply nested extraction paths is possible.

* Add more strict typing to grpc handler

In addtion to the stricter typing, also add a check to ensure that the
service clients being constructed are actually constructable. If not
throw an Error for the user.

Also remove the custom type definition for util.promisify and use the
type defintion from Node itself.