Make documentation more precise and correct the order of steps.
Before the CLI command can be executed, the `require: -ts-node...` should be in the .yaml file, thus I've moved it above the CLI step.

The API here only describes how to generate Typescript types for the entire schema. From this API it is clear that that is the only option available, and the description should be clear on that part. 
If, however, it is possible to do achieve more, then please clarify that in the documentation.