Fix CodeSandbox on landing page
The CodeSandbox on Mesh website was broken because we’re getting a cache hit from the website-router which strips the search params. Unlucky, because we’re using search params to pass the CodeSandbox id to the middleman iframe, and we need the middleman iframe because CodeSandbox scrolls to itself on load :)

I changed the iframe and the sandbox component to use postMessage instead.