Fix CodeSandbox on landing page (#8768)
* Fix CodeSandbox on landing page

The CodeSandbox on Mesh website was broken because we’re getting a cache hit from the website-router which strips the search params. Unlucky, because we’re using search params to pass the CodeSandbox id to the middleman iframe, and we need the middleman iframe because CodeSandbox scrolls to itself on load :)

I changed the iframe and the sandbox component to use postMessage instead.

* Add missing ;

* Harden against XSS (thanks @coderabbitai)

* Stop passing * as target origin

* 📝 Add docstrings to `fix-sandbox-on-landing` (#8772)

Docstrings generation was requested by @hasparus.

* https://github.com/ardatan/graphql-mesh/pull/8768#issuecomment-3298962637

The following files were modified:

* `website/src/components/examples-sandbox.tsx`

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>